Fix some -Wfloat-conversion warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31690 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/pdf/pdfwritabledataset.cpp
@@ -253,12 +253,12 @@ OGRErr PDFWritableVectorDataset::SyncToDisk()
     if (dfRatio < 1)
     {
         nWidth = 1024;
-        nHeight = nWidth * dfRatio;
+        nHeight = static_cast<int>(nWidth * dfRatio);
     }
     else
     {
         nHeight = 1024;
-        nWidth = nHeight / dfRatio;
+        nWidth = static_cast<int>(nHeight / dfRatio);
     }
 
     GDALDataset* poSrcDS = MEMDataset::Create( "MEM:::", nWidth, nHeight, 0, GDT_Byte, NULL );
